feat: add Shorts durations + new niches to UI, auto-sync niche/format/duration
- Duration selector now has 30s / 45s / 1min options under Shorts group
- Added history_short and face_reconstruction niches to dropdown
- JS auto-syncs duration + format when niche is changed
- app.py: float() instead of int() for duration (supports 0.5, 0.75)
- script_writer: int() cast on word count to avoid float display

+++ script_writer.py
@@ -314,7 +314,7 @@ def generate_script(topic: str, duration_minutes: int = None, niche: str = "fact
 Niche: {niche_info['label']}
 Tone: {niche_info['tone']}
 Style guidelines: {niche_info['style']}
-Target length: {duration_minutes} minutes (~{duration_minutes * 150} words of spoken narration)
+Target length: {duration_minutes} minutes (~{int(duration_minutes * 150)} words of spoken narration)
 {hook_instruction}
 
 VOICE AND STYLE — NON-NEGOTIABLE:
